+++ src/lib/find-parent-form.js
@@ -1,0 +1,15 @@
+'use strict';
+
+function findParentForm(element) {
+  var parentNode = element.parentNode;
+
+  if (!parentNode || parentNode.nodeName === 'FORM') {
+    return parentNode;
+  }
+
+  return findParentForm(parentNode);
+}
+
+module.exports = {
+  findParentForm: findParentForm
+};